refactor: bootstrap launcher
